Refresh deployment before rolling upgrade
Before, we fetched deployment objects only once. The problem is: if
the deployment manifest was changed, we could have re-applied an older
version of the manifests. Fortunately, it seems that Kubernetes was
aware of that and prevented the deployment update from happening.

What we do now is we systematically GET the deployment before updating
it to trigger the rolling upgrade. This should prevent the issue
(unless someone updates the manifests after the GET and before the
PATH, which is extremely unlikely :)).